[Wsign-compare]
[Changes]
- Enable -Wsign-compare for ARC CLANG.

+++ lib/src/bricks/impl/mli_krn_dotprod_vdsp.h
@@ -1,5 +1,5 @@
 /*
-* Copyright 2020, Synopsys, Inc.
+* Copyright 2020-2021, Synopsys, Inc.
 * All rights reserved.
 *
 * This source code is licensed under the BSD-3-Clause license found in
@@ -331,7 +331,7 @@ static MLI_FORCE_INLINE vNx2short_t make_vindex(
         int unroll = 1,
         int in_unroll_step = 0) {
     vNx2short_t vindex;
-    MLI_ASSERT(width * height * unroll <= (sizeof(vNx2short_t) / sizeof(short)));
+    MLI_ASSERT(width * height * unroll <= (int)(sizeof(vNx2short_t) / sizeof(short)));
     int idx = 0;
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wpass-failed"
